broke down the test suite in smaller and namageable modules
